+++ include/emulator/class.h
