+++ include/emulator/ram.h
