+++ include/emulator/system.h
