+++ include/emulator/xlHeap.h
