+++ src/emulator/rom.c
