+++ src/emulator/tree.c
