+++ src/emulator/xlHeap.c
