+++ src/emulator/xlText.c
