+++ src/revolution/os/__start.c
@@ -17,6 +17,26 @@ extern __bss_init __bss_init_info[];
 void* memcpy(void* dst, void* src, unsigned int len);
 void __flush_cache(void* src, unsigned int len);
 void* memset(void* dst, int fill, unsigned int len);
+void OSResetSystem(int, int, int);
+
+extern u8 Debug_BBA;
+
+void __check_pad3(void) {
+    unsigned short* t = (unsigned short*)0x800030E4;
+    if ((*t & 0xEEF) != 0xEEF) {
+        return;
+    }
+
+    OSResetSystem(0x0, 0x0, 0x0);
+}
+
+void __set_debug_bba(void) {
+    Debug_BBA = 1;
+}
+
+u8 __get_debug_bba(void) {
+    return Debug_BBA;
+}
 
 #ifdef NON_MATCHING
 void __init_data(void) {
